System core's fsinit patch now includes a fix for the /system/bin/auditctl permissions (stickybit wasn't set) so it has UID root, GID Audit, and will run as root (required for auditctl startup checks).
The /etc/audit/auditd.rules file now has a GID of audit and permissions of
660 so that it's read/write for members of the audit group. This eventually
will allow an API to modify the auditd.rules file (checking for the
android.permission.AUDIT permission) and then call auditctl to reload the
rules. Auditctl can also be used from an android.permission.AUDIT enabled
APK to check the status of audit.